Move dataset handling out of TimelineView
